+++ test/unit/tools/bigfiles.rb
@@ -6,7 +6,7 @@ module Test
         def bigfiles_expected_args
           '--glob ' \
           "'{src,app,lib,test,spec,feature}/**/{Rakefile,*." \
-          "{rb,swift,cpp,c,java,py,clj,cljs}}'"
+          "{rb,swift,cpp,c,java,py,clj,cljs,scala,js}}'"
         end
 
         def expect_bigfiles_run(quality_checker)