+++ test/unit/tools/cane.rb
@@ -4,7 +4,7 @@ module Test
       # Test for the 'cane' tool within the quality gem
       module Cane
         def cane_expected_args
-          "-f '{src,app,lib,test,spec,feature}/**/{Rakefile,*.{rb}}'"
+          "-f '{src,app,lib,test,spec,feature}/**/{Rakefile,*.{rb,rake}}'"
         end
 
         def expect_cane_run(quality_checker)