+++ ChangeLog.md
@@ -52,3 +52,6 @@
 ## Changes since Quality 6.0.0
 ### Add Rakefile to default source files (configurable via extra_files parameter)
 ### Add basic support for Scala and JavaScript via language-independent tools like bigfiles and punchlist
+
+## Changes since Quality 7.0.0
+### Add support for .rake files as Ruby language